requirements.txt
torch~=2.4.0
transformers~=4.44.2

pytesseract~=0.3.13
pillow~=10.4.0